third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-window-name-navigator-expected.txt

CONSOLE MESSAGE: line 6: iframe1: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame.
CONSOLE MESSAGE: line 12: FAIL
CONSOLE MESSAGE: line 4: FAIL
 
